Putting eval in A2C and A3C algos
- Now we can run a test batch
- Adding jupyter notebook
- Refactoring PPO loss functions
