+++ PPO/ppo.py
@@ -12,7 +12,8 @@ from .critic import Critic
 from .actor import Actor
 from utils.networks import tfSummary
 from utils.stats import gather_stats
-from keras import backend as K
+
+from .ppo_loss import proximal_policy_optimization_loss 
 
 class PPO:
     """ Actor-Critic Main Algorithm
@@ -290,16 +291,3 @@ class PPO:
             self.gradient_steps += 1
 
         return rewards, actor_losses, critic_losses
-
-
-def exponential_average(old, new, b1):
-    return old * b1 + (1-b1) * new
-
-
-def proximal_policy_optimization_loss(advantage, old_prediction, loss_clipping, entropy_loss):
-    def loss(y_true, y_pred):
-        prob = K.sum(y_true * y_pred, axis=-1)
-        old_prob = K.sum(y_true * old_prediction, axis=-1)
-        r = prob/(old_prob + 1e-10)
-        return -K.mean(K.minimum(r * advantage, K.clip(r, min_value=1 - loss_clipping, max_value=1 + loss_clipping) * advantage) + entropy_loss * -(prob * K.log(prob + 1e-10)))
-    return loss
